Update sprint file Valesca + UGRID read
I have updated the poetry installation with python 3.9 as scipy could not work with python 3.10. In addition i have added the changes Valesca could not commit on the sprint file and renamed it to one single file to distinguish it from the UGRID translation (also added). For the ugrid translation the topobathy file needed to be added.